[#1554] Refactor CustomDomain model
To be like Mike. Err... I mean Customer.
