+++ apps/api/v2/models/custom_domain.rb
@@ -34,93 +34,6 @@ require 'public_suffix'
 module V2
   class CustomDomain < Familia::Horreum
 
-    unless defined?(MAX_SUBDOMAIN_DEPTH)
-      MAX_SUBDOMAIN_DEPTH = 10 # e.g., a.b.c.d.e.f.g.h.i.j.example.com
-      MAX_TOTAL_LENGTH    = 253   # RFC 1034 section 3.1
-    end
-
-    prefix :customdomain
-
-    feature :safe_dump
-
-    identifier :derive_id
-
-    # NOTE: The redis key used by older models for values is simply
-    # "onetime:customdomain". We'll want to rename those at some point.
-    class_sorted_set :values
-    class_hashkey :display_domains
-    class_hashkey :owners
-
-    field :display_domain
-    field :custid
-    field :domainid
-    field :base_domain
-    field :subdomain
-    field :trd
-    field :tld
-    field :sld
-    field :txt_validation_host
-    field :txt_validation_value
-    field :status
-    field :vhost
-    field :verified # the txt record matches?
-    field :resolving # there's a valid A or CNAME record?
-    field :created
-    field :updated
-    field :_original_value
-
-    hashkey :brand
-    hashkey :logo # image fields need a corresponding v2 route and logic class
-    hashkey :icon
-
-    @txt_validation_prefix = '_onetime-challenge'.freeze
-
-    @safe_dump_fields = [
-      { identifier: ->(obj) { obj.identifier } },
-      :domainid,
-      :display_domain,
-      :custid,
-      :base_domain,
-      :subdomain,
-      :trd,
-      :tld,
-      :sld,
-      { is_apex: ->(obj) { obj.apex? } },
-      :_original_value,
-      :txt_validation_host,
-      :txt_validation_value,
-      { brand: ->(obj) { obj.brand.hgetall } },
-      # NOTE: We don't serialize images here
-      :status,
-      { vhost: ->(obj) { obj.parse_vhost } },
-      :verified,
-      :created,
-      :updated,
-    ].freeze
-
-    def init
-      @domainid = identifier
-
-      # Display domain and cust should already be set and accessible
-      # via accessor methods so we should see a valid identifier logged.
-      OT.ld "[CustomDomain.init] #{display_domain} id:#{domainid}"
-
-      # Will raise PublicSuffix::DomainInvalid if invalid domain
-      ps_domain = PublicSuffix.parse(display_domain, default_rule: nil)
-
-      # Store the individual domain parts that PublicSuffix parsed out
-      @base_domain = ps_domain.domain.to_s
-      @subdomain   = ps_domain.subdomain.to_s
-      @trd         = ps_domain.trd.to_s
-      @tld         = ps_domain.tld.to_s
-      @sld         = ps_domain.sld.to_s
-
-      # Don't call generate_txt_validation_record here otherwise we'll
-      # create a new validation record every time we instantiate a
-      # custom domain object. Instead, we'll call it when we're ready
-      # to verify the domain.
-    end
-
     # Generate a unique identifier for this customer's custom domain.
     #
     # From a customer's perspective, the display_domain is what they see
@@ -381,214 +294,8 @@ module V2
       verification_state == :verified
     end
 
-    module ClassMethods
-      attr_reader :db, :values, :owners, :txt_validation_prefix
-
-      # Creates a new custom domain record
-      #
-      # This method:
-      # 1. Validates and parses the input domain
-      # 2. Checks for duplicates
-      # 3. Saves the domain and updates related records atomically
-      #
-      # @param input [String] The domain name to create
-      # @param custid [String] The customer ID to associate with
-      # @return [V2::CustomDomain] The created custom domain
-      # @raise [Onetime::Problem] If domain is invalid or already exists
-      #
-      # More Info:
-      # We need a minimum of a domain and customer id to create a custom
-      # domain -- or more specifically, a custom domain indentifier. We
-      # allow instantiating a custom domain without a customer id, but
-      # instead raise a fuss if we try to save it later without one.
-      #
-      # See CustomDomain.base_domain and display_domain for details on
-      # the difference between display domain and base domain.
-      #
-      # NOTE: Interally within this class, we try not to use the
-      # unqualified term "domain" on its own since there's so much
-      # room for confusion.
-      #
-      def create(input, custid)
-        obj = parse(input, custid)
-
-        redis.watch(obj.rediskey) do
-          if obj.exists?
-            redis.unwatch
-            raise Onetime::Problem, 'Duplicate domain for customer'
-          end
-
-          redis.multi do |_multi|
-            obj.generate_txt_validation_record
-            obj.save
-            # Create minimal customer instance for Redis key
-            cust = V2::Customer.new(custid: custid)
-            cust.add_custom_domain(obj)
-            # Add to global values set
-            add(obj)
-          end
-        end
-
-        obj  # Return the created object
-      rescue Redis::BaseError => ex
-        OT.le "[CustomDomain.create] Redis error: #{ex.message}"
-        raise Onetime::Problem, 'Unable to create custom domain'
-      end
-
-      # Returns a new V2::CustomDomain object (without saving it).
-      #
-      # @param input [String] The domain name to parse
-      # @param custid [String] Customer ID associated with the domain
-      #
-      # @return [V2::CustomDomain]
-      #
-      # @raise [PublicSuffix::DomainInvalid] If domain is invalid
-      # @raise [PublicSuffix::DomainNotAllowed] If domain is not allowed
-      # @raise [PublicSuffix::Error] For other PublicSuffix errors
-      # @raise [Onetime::Problem] If domain exceeds MAX_SUBDOMAIN_DEPTH or MAX_TOTAL_LENGTH
-      #
-      def parse(input, custid)
-        raise Onetime::Problem, 'Customer ID required' if custid.to_s.empty?
-
-        segments = input.to_s.split('.').reject(&:empty?)
-        raise Onetime::Problem, 'Invalid domain format' if segments.empty?
-
-        if segments.length > MAX_SUBDOMAIN_DEPTH
-          raise Onetime::Problem, "Domain too deep (max: #{MAX_SUBDOMAIN_DEPTH})"
-        end
-
-        if input.length > MAX_TOTAL_LENGTH
-          raise Onetime::Problem, "Domain too long (max: #{MAX_TOTAL_LENGTH})"
-        end
-
-        display_domain      = self.display_domain(input)
-        obj                 = new(display_domain, custid)
-        obj._original_value = input
-        obj
-      end
-
-      # Takes the given input domain and returns the base domain,
-      # the one that the zone record would be created for. So
-      # froogle.com, www.froogle.com, subdir.www.froogle.com would
-      # all return froogle.com here.
-      #
-      # Another way to think about it, the TXT record we ask the user
-      # to create will be created on the base domain. So if we have
-      # www.froogle.com, we'll create the TXT record on froogle.com,
-      # like this: `_onetime-challenge-domainid.froogle.com`. This is
-      # distinct from the domain we ask the user to create an A
-      # record for, which is www.froogle.com. We also call this the
-      # display domain.
-      #
-      # Returns either a string or nil if invalid
-      def base_domain(input)
-        # We don't need to fuss with empty stripping spaces, prefixes,
-        # etc because PublicSuffix does that for us.
-        PublicSuffix.domain(input, default_rule: nil)
-      rescue PublicSuffix::DomainInvalid => ex
-        OT.le "[CustomDomain.base_domain] #{ex.message} for `#{input}`"
-        nil
-      end
-
-      # Takes the given input domain and returns the display domain,
-      # the one that we ask the user to create an A record for. So
-      # subdir.www.froogle.com would return subdir.www.froogle.com here;
-      # www.froogle.com would return www.froogle.com; and froogle.com
-      # would return froogle.com.
-      #
-      def display_domain(input)
-        ps_domain = PublicSuffix.parse(input, default_rule: nil)
-        ps_domain.subdomain || ps_domain.domain
-      rescue PublicSuffix::Error => ex
-        OT.le "[CustomDomain.parse] #{ex.message} for `#{input}`"
-        raise Onetime::Problem, ex.message
-      end
-
-      # Returns boolean, whether the domain is a valid public suffix
-      # which checks without actually parsing it.
-      def valid?(input)
-        PublicSuffix.valid?(input, default_rule: nil)
-      end
-
-      def default_domain?(input)
-        display_domain = V2::CustomDomain.display_domain(input)
-        site_host      = OT.conf.dig(:site, :host)
-        OT.ld "[CustomDomain.default_domain?] #{display_domain} == #{site_host}"
-        display_domain.eql?(site_host)
-      rescue PublicSuffix::Error => ex
-        OT.le "[CustomDomain.default_domain?] #{ex.message} for `#{input}"
-        false
-      end
-
-      # Simply instatiates a new CustomDomain object and checks if it exists.
-      def exists?(input, custid)
-        # The `parse`` method instantiates a new CustomDomain object but does
-        # not save it to Redis. We do that here to piggyback on the inital
-        # validation and parsing. We use the derived identifier to load
-        # the object from Redis using
-        obj = parse(input, custid)
-        OT.ld "[CustomDomain.exists?] Got #{obj.identifier} #{obj.display_domain} #{obj.custid}"
-        obj.exists?
-      rescue Onetime::Problem => ex
-        OT.le "[CustomDomain.exists?] #{ex.message}"
-        false
-      end
-
-      def add(fobj)
-        values.add OT.now.to_i, fobj.to_s # created time, identifier
-        display_domains.put fobj.display_domain, fobj.identifier
-        owners.put fobj.to_s, fobj.custid  # domainid => customer id
-      end
-
-      def rem(fobj)
-        values.remove fobj.to_s
-        display_domains.remove fobj.display_domain
-        owners.remove fobj.to_s
-      end
-
-      def all
-        # Load all instances from the sorted set. No need
-        # to involve the owners HashKey here.
-        values.revrangeraw(0, -1).collect { |identifier| from_identifier(identifier) }
-      end
-
-      def recent(duration = 48.hours)
-        spoint = OT.now.to_i-duration
-        epoint = OT.now.to_i
-        values.rangebyscoreraw(spoint, epoint).collect { |identifier| load(identifier) }
-      end
-
-      # Implement a load method for CustomDomain to make sure the
-      # correct derived ID is used as the key.
-      def load(display_domain, custid)
-        custom_domain = parse(display_domain, custid).tap do |obj|
-          OT.ld "[CustomDomain.load] Got #{obj.identifier} #{obj.display_domain} #{obj.custid}"
-          raise Onetime::RecordNotFound, "Domain not found #{obj.display_domain}" unless obj.exists?
-        end
-
-        # Continue with the built-in `load` from Familia.
-        super(custom_domain.identifier)
-      end
-
-      # Load a custom domain by display domain only. Used during requests
-      # after determining the domain strategy is :custom.
-      #
-      # @param display_domain [String] The display domain to load
-      # @return [V2::CustomDomain, nil] The custom domain record or nil if not found
-      def from_display_domain(display_domain)
-        # Get the domain ID from the display_domains hash
-        domain_id = display_domains.get(display_domain)
-        return nil unless domain_id
-
-        # Load the record using the domain ID
-        begin
-          from_identifier(domain_id)
-        rescue Onetime::RecordNotFound
-          nil
-        end
-      end
-    end
-
-    extend ClassMethods
   end
 end
+
+require_relative 'custom_domain/definition'
+require_relative 'custom_domain/class_methods'
